search working on Projects page, also updated utilities.js, small CSS change for search field
